blasso.s3 {monomvn}	R Documentation

Summarizing Bayesian Lasso Output

Description

Summarizing, printing, and plotting the contents of a "blasso"-class object containing samples from the posterior distribution of a Bayesian lasso model

Usage

## S3 method for class 'blasso'
print(x, ...)
## S3 method for class 'blasso'
summary(object, burnin = 0, ...)
## S3 method for class 'blasso'
plot(x, which=c("coef", "s2", "lambda2", "gamma",
    "tau2i","omega2", "nu", "m", "pi"), subset = NULL, burnin = 0,
    ... )
## S3 method for class 'summary.blasso'
print(x, ...)

Arguments

object
a "blasso"-class object that must be named object for the generic methods summary.blasso

x
a "blasso"-class object that must be named x for the generic printing and plotting methods print.summary.blasso and plot.blasso

subset
a vector of indicies that can be used to specify the a subset of the columns of tau2i or omega2 that are plotted as boxplots in order to reduce clutter

burnin
number of burn-in rounds to discard before reporting summaries and making plots. Must be non-negative and less than x$T

which
indicates the parameter whose characteristics should be plotted; does not apply to the summary

...
passed to print.blasso, or plot.default

Details

print.blasso prints the call followed by a brief summary of the MCMC run and a suggestion to try the summary and plot commands.

plot.blasso uses an appropriate plot command on the list entries of the "blasso"-class object thus visually summarizing the samples from the posterior distribution of each parameter in the model depending on the which argument supplied.

summary.blasso uses the summary command on the list entries of the "blasso"-class object thus summarizing the samples from the posterior distribution of each parameter in the model.

print.summary.monomvn calls print.blasso on the object and then prints the result of summary.blasso

Value

summary.blasso returns a "summary.blasso"-class object, which is a list containing (a subset of) the items below. The other functions do not return values.

B
a copy of the input argument thin

T
total number of MCMC samples to be collected from x$T

thin
number of MCMC samples to skip before a sample is collected (via thinning) from x$T

coef
a joint summary of x$mu and the columns of x$beta, the regression coefficients

s2
a summary of x$s2, the variance parameter

lambda2
a summary of x$lambda2, the penalty parameter, when lasso or ridge regression is active

lambda2
a summary of x$gamma, when the NG extensions to the lasso are used

tau2i
a summary of the columns of the latent x$tau2i parameters when lasso is active

omega2
a summary of the columns of the latent x$omega2 parameters when Student-t errors are active

nu
a summary of x$nu, the degrees of freedom parameter, when the Student-t model is active

bn0
the estimated posterior probability that the individual components of the regression coefficients beta is nonzero

m
a summary the model order x$m: the number of non-zero regression coefficients beta

pi
the estimated Binomial proportion in the prior for the model order when 2-vector input is provided for mprior

Author(s)

Robert B. Gramacy rbgramacy@chicagobooth.edu

References

http://bobby.gramacy.com/r_packages/monomvn

See Also

blasso

[Package monomvn version 1.9-6 Index]