monomvn-package {monomvn}	R Documentation

Estimation for Multivariate Normal and Student-t Data with Monotone Missingness

Description

Estimation of multivariate normal and student-t data of arbitrary dimension where the pattern of missing data is monotone. Through the use of parsimonious/shrinkage regressions (plsr, pcr, lasso, ridge, etc.), where standard regressions fail, the package can handle a nearly arbitrary amount of missing data. The current version supports maximum likelihood inference and a full Bayesian approach employing scale-mixtures for Gibbs sampling. Monotone data augmentation extends this Bayesian approach to arbitrary missingness patterns. A fully functional standalone interface to the Bayesian lasso (from Park & Casella), the Normal-Gamma (from Griffin & Brown), Horseshoe (from Carvalho, Polson, & Scott), and ridge regression with model selection via Reversible Jump, and student-t errors (from Geweke) is also provided

Details

For a fuller overview including a complete list of functions, demos and vignettes, please use help(package="monomvn").

Author(s)

Robert B. Gramacy rbgramacy@chicagobooth.edu

Maintainer: Robert B. Gramacy rbgramacy@chicagobooth.edu

References

Robert B. Gramacy, Joo Hee Lee and Ricardo Silva (2008). On estimating covariances between many assets with histories of highly variable length.
Preprint available on arXiv:0710.5837: http://arxiv.org/abs/0710.5837

http://bobby.gramacy.com/r_packages/monomvn

See Also

monomvn, the now defunct norm package, mvnmle

[Package monomvn version 1.9-6 Index]