monomvn.s3 {monomvn}	R Documentation

Summarizing monomvn output

Description

Summarizing, printing, and plotting the contents of a "monomvn"-class object

Usage

## S3 method for class 'monomvn'
summary(object, Si = FALSE, ...)
## S3 method for class 'summary.monomvn'
print(x, ...)
## S3 method for class 'summary.monomvn'
plot(x, gt0 = FALSE, main = NULL,
     xlab = "number of zeros", ...)

Arguments

object
a "monomvn"-class object that must be named object for the generic methods summary.monomvn

x
a "monomvn"-class object that must be named x for generic printing and plotting via print.summary.monomvn and plot.summary.monomvn

Si
boolean indicating whether object$S should be inverted and inspected for zeros within summary.monomvn, indicating pairwise independence; default is FALSE

gt0
boolean indicating whether the histograms in plot.summary.monomvn should exclude columns of object$S or Si without any zero entries

main
optional text to be added to the main title of the histograms produced by the generic plot.summary.monomvn

xlab
label for the x-axes of the histograms produced by plot.summary.monomvn; otherwise default automatically-generated text is used

...
passed to print.monomvn, or plot.default

Details

These functions work on the output from both monomvn and bmonomvn.

print.monomvn prints the call followed by a summary of the regression method used at each iteration of the algorithm. It also indicates how many completely observed features (columns) there were in the data. For non-least-squares regressions (i.e., plsr, lars and lm.ridge methods) and indication of the method used for selecting the number of components (i.e., CV, LOO, etc., or none) is provided

summary.monomvn summarizes information about the number of zeros in the estimated covariance matrix object$S and its inverse

print.summary.monomvn calls print.monomvn on the object and then prints the result of summary.monomvn

plot.summary.monomvn makes histograms of the number of zeros in the columns of object$S and its inverse

Value

summary.monomvn returns a "summary.monomvn"-class object, which is a list containing (a subset of) the items below. The other functions do not return values.

obj
the "monomvn"-class object

marg
the proportion of zeros in object$S

S0
a vector containing the number of zeros in each column of object$S

cond
if input Si = TRUE this field contains the proportion of zeros in the inverse of object$S

Si0
if input Si = TRUE this field contains a vector with the number of zeros in each column of the inverse of object$S

Note

There is one further S3 function for "monomvn"-class objects that has its own help file: plot.monomvn

Author(s)

Robert B. Gramacy rbgramacy@chicagobooth.edu

References

http://bobby.gramacy.com/r_packages/monomvn

See Also

bmonomvn, monomvn, plot.monomvn

[Package monomvn version 1.9-6 Index]