monomvn.solve.QP {monomvn}	R Documentation

Solve a Quadratic Program

Description

Solve a Quadratic Program specified by a QP object using the covariance matrix and mean vector specified

Usage

monomvn.solve.QP(S, QP, mu = NULL)

Arguments

S
a positive-definite covariance matrix whose dimensions agree with the Quadratic Program, e.g., nrow(QP$Amat)

QP
a Quadratic Programming object like one that can be generated automatically by default.QP

mu
an mean vector with length(mu) = nrow(QP$Amat) that is required if QP$dmu == TRUE or QP$mu.constr[1] != 0

Details

The protocol executed by this function is identical to the one used on samples of S and mu obtained in bmonomvn when a Quadratic Program is specified through the QP argument. For more details on the specification of the Quadratic Program implied by a QP object, please see default.QP and the examples therein

Value

The output is a vector whose length agrees with the dimension of S, describing the solution to the Quadratic Program given

Author(s)

Robert B. Gramacy rbgramacy@chicagobooth.edu

See Also

default.QP, bmonomvn, and solve.QP in the quadprog package

[Package monomvn version 1.9-6 Index]