rwish {monomvn}	R Documentation

Draw from the Wishart Distribution

Description

Random generation from the Wishart distribution.

Usage

   rwish(v, S)

Arguments

v
degrees of freedom (scalar)

S
inverse scale matrix (p x p)

Details

The mean of a Wishart random variable with v degrees of freedom and inverse scale matrix S is vS.

Value

Returns generates one random draw from the distribution which is a matrix with the same dimensions as S

References

This was copied from the MCMCpack package

Examples

draw <- rwish(3, matrix(c(1,.3,.3,1),2,2))

[Package monomvn version 1.9-6 Index]